fix(platform): 第二轮全仓深审——client 并发三连、消息面租户隔离、LB/trace/4xx 分类等 17 项修复
P1：
- client CLI 并发三连：registry watch 回调锁外调 LB（connect 换 LB → UAF）
  移入锁内；心跳/用户线程并发 reconnect() → reconnect_mutex_ 串行化；
  AIQueryClient stub 裸指针并发重建 → shared_ptr + 锁内快照
- agent_service 消息面无租户隔离：任意认证用户可读走/注入他租户消息
  → 队列属主表（注册者），Receive/Listen 校验属主；SendMessage 打
  x-sender-owner 头（proto 契约不动）

P2：
- LB markEndpointStatus 对 Random/LeastConnections/LeastResponseTime
  三策略无效 → select 统一内部健康记录优先
- a2a_client 非 2xx 全归 transport 可重试（4xx 风暴打 OPEN 熔断）→
  4xx 除 429 抛 HTTP protocol error，不重试不入熔断
- trace_context::endSpan 持栈引用调 exporter（重入错乱/析构逃逸
  terminate）→ pop 前置 + exporter 异常吞并
- redis_client incrby 变参 %ld 传 int64_t（LLP64 UB）→ %lld
- Heartbeat 对任意 agent_id 无条件 upsert durable healthy 行（绕过
  ADMIN 门槛）→ 仅本进程注册过的 id 刷新

P3：sharing expiry_days 溢出上限、randomHex 移入 try×3、cost_tracker
投影失败告警、client 信号处理器 async-signal-safe 化、README 撤销
口径与 agent-integration-guide ADMIN 首注抢占注记

风格：P-tag 注释归一、重复注释压缩、registry 头注释去重
验证：门禁 37/37、MCP 2/2、网关 116/116、typecheck、E2E 25/25、redline

+++ common/src/redis_client.cpp
@@ -362,7 +362,8 @@ bool RedisClient::incrby(const std::string& key, int64_t increment,
     if (!ensureConnected()) return false;
 
     auto* reply = static_cast<redisReply*>(
-        redisCommand(ctx_, "INCRBY %s %ld", key.c_str(), increment));
+        redisCommand(ctx_, "INCRBY %s %lld", key.c_str(),
+                     static_cast<long long>(increment)));
     if (!reply) return false;
 
     if (reply->type == REDIS_REPLY_INTEGER) {
